fixed typos and debug statements and renamed others to utils
