+++ wcosa/parsers/board_parser.py
@@ -4,7 +4,7 @@ Parses the boards.txt file and gathers information about the current board
 
 import json
 
-from others import helper
+from utils import helper
 
 
 def create_boards_tree(board_file_path, new_board_path):
