+++ wcosa/parsers/platform_parser.py
@@ -4,7 +4,7 @@ Parses the platform.txt file and gathers information about the current platform
 
 import json
 
-from others import helper
+from utils import helper
 
 
 def get_raw_flags(lines, identifier, include_extra):
